# Improve formula performance