--- a/src/BlazorDatasheet.Core/Data/Cells/CellStore.Validation.cs
+++ b/src/BlazorDatasheet.Core/Data/Cells/CellStore.Validation.cs
@@ -22,7 +22,7 @@ public partial class CellStore
             _validStore.Set(row, col, result.IsValid);
         }
 
-        Sheet.MarkDirty(cellsAffected);
+        Sheet.MarkDirty(region);
     }
 
     public bool IsValid(int row, int col)