--- a/src/BlazorDatasheet.Core/Data/Sheet.cs
+++ b/src/BlazorDatasheet.Core/Data/Sheet.cs
@@ -330,15 +330,6 @@ public class Sheet
     }
 
     /// <summary>
-    /// Mark the cells specified by positions dirty.
-    /// </summary>
-    /// <param name="positions"></param>
-    internal void MarkDirty(IEnumerable<CellPosition> positions)
-    {
-        MarkDirty(positions.Select(p => new Region(p.row, p.col)));
-    }
-
-    /// <summary>
     /// Marks the cell as dirty and requiring re-render
     /// </summary>
     /// <param name="row"></param>
@@ -358,7 +349,9 @@ public class Sheet
         if (intersection == null)
             return;
 
-        MarkDirty(new List<IRegion>() { intersection });
+        _dirtyRegions.Add(intersection, true);
+        if (!_isBatchingChanges)
+            EmitSheetDirty();
     }
 
     /// <summary>
@@ -377,13 +370,16 @@ public class Sheet
         }
 
         if (!_isBatchingChanges)
+            EmitSheetDirty();
+    }
+
+    private void EmitSheetDirty()
+    {
+        SheetDirty?.Invoke(this, new()
         {
-            SheetDirty?.Invoke(this, new DirtySheetEventArgs()
-            {
-                DirtyRegions = _dirtyRegions,
-            });
-            _dirtyRegions.Clear();
-        }
+            DirtyRegions = _dirtyRegions
+        });
+        _dirtyRegions.Clear();
     }
 
     private int _batchRequestNo;
@@ -414,9 +410,12 @@ public class Sheet
 
         var beforeArgs = new BeforeRangeSortEventArgs(region, sortOptions);
         BeforeRangeSort?.Invoke(this, beforeArgs);
+        
         var cmd = new SortRangeCommand(region, sortOptions);
+        
         if (!beforeArgs.Cancel)
             Commands.ExecuteCommand(cmd);
+        
         var afterArgs = new RangeSortedEventArgs(region, cmd.SortedRegion, cmd.OldIndices);
         RangeSorted?.Invoke(this, afterArgs);
     }
@@ -436,15 +435,9 @@ public class Sheet
         // Checks for batching changes here, because the cells changed event
         // may start batching more dirty changes again from subscribers of that event.
         if (_dirtyRegions.Any() && _isBatchingChanges)
-        {
-            SheetDirty?.Invoke(this, new DirtySheetEventArgs()
-            {
-                DirtyRegions = _dirtyRegions,
-            });
-        }
+            EmitSheetDirty();
 
         _isBatchingChanges = false;
-        _dirtyRegions.Clear();
     }
 
     /// <summary>