--- a/test/BlazorDatasheet.Test/Commands/SetCellValuesCommandTests.cs
+++ b/test/BlazorDatasheet.Test/Commands/SetCellValuesCommandTests.cs
@@ -12,12 +12,12 @@ public class SetCellValuesCommandTests
     public void Set_Cell_Values_Respects_Cell_Type()
     {
         var sheet = new Sheet(10, 10);
-        sheet.Commands.ExecuteCommand(new SetCellValuesCommand(0, 0, [["2020-09-09"]]));
+        sheet.Cells.SetValues(0, 0, [["2020-09-09"]]);
         // ensure the conversion happens without setting type first
         sheet.Cells.GetValue(0, 0).Should().BeOfType<DateTime>();
         sheet.Cells.SetType(0, 0, "text");
         // now since the type is "text" the conversion should not happen
-        sheet.Commands.ExecuteCommand(new SetCellValuesCommand(0, 0, [["2020-09-09"]]));
+        sheet.Cells.SetValues(0, 0, [["2020-09-09"]]);
         sheet.Cells.GetValue(0, 0).Should().BeOfType<string>();
     }