Uses some evaluated formula caching and better dependency management of single cell references.

Using similar code to #198:

```csharp
_sheet.BatchUpdates();
for (int i = 0; i < 200; i++)
{
    for (int j = 0; j < 200; j++)
    {
        var cell = RangeText.ToCellText(i + 1, j);
        _sheet.Cells.SetFormula(i, j, $"={cell}+1");
    }
}

_sheet.EndBatchUpdates();
```

We have the following results prior to this PR:

| Method         | Mean    | Error   | StdDev  | Allocated |
|--------------- |--------:|--------:|--------:|----------:|
| Set200x200Formula | 11.33 s | 0.150 s | 0.133 s |   1.46 GB |

And then after:

| Method         | Mean     | Error    | StdDev   | Allocated |
|--------------- |---------:|---------:|---------:|----------:|
| Set200x200Formula | 861.5 ms | 22.00 ms | 63.47 ms |   1.03 GB |

We can likely still get a lot better than this, especially memory allocations.